% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenueCES.R
\name{revenueCES}
\alias{revenueCES}
\title{Two-step estimation of revenue CES}
\usage{
revenueCES(
  Y,
  X,
  V,
  P_V,
  idvar,
  timevar,
  v_bounds = FALSE,
  v = NULL,
  mu = NULL,
  boot = 100,
  opt = "optim",
  theta_0 = NULL,
  cluster = NULL,
  control = NULL
)
}
\arguments{
\item{Y}{A vector of firms' revenue over time.}

\item{X}{A vector of state / dynamic inputs (e.g. capital).}

\item{V}{A vector of variable inputs (e.g. material inputs).}

\item{P_V}{A vector of firm- or industry-level input prices, ranked in the
same way as the corresponding inputs in \code{V}.}

\item{idvar}{A vector indexing which firm each row of the vectors corresponds to.}

\item{timevar}{A vector indexing which time period each row of the vectors corresponds to}

\item{v_bounds}{A Boolean indicating whether bounds on the returns-to-scale parameter
v should be computed. Default = FALSE.}

\item{v}{An optional number in \eqn{]0,\infty[} to which the returns-to-scale parameter
should be restricted. Constant returns to scale correspond to \code{v} = 1. Default = NULL.}

\item{mu}{A Boolean whether firm-level markups (with error term) should be computed and reported.
Requires that \code{v} is specified. Default = FALSE.}

\item{boot}{Number of bootstrap repetitions for calculating standard errors. Default = 100.}

\item{opt}{A string with the optimization algorithm to be used during the GMM estimation.
Options are \link[stats]{optim}, \link[Rsolnp]{solnp} and \link[DEoptim]{DEoptim}. Default = 'optim'.}

\item{theta_0}{An optional named vector of starting values for the parameters,
where the names are 'beta_L', 'beta_M', 'sigma'. Default = NULL.}

\item{cluster}{An optional object of class "SOKcluster" or "cluster" for parallelization of the bootstrap.
See \link[paralell]{makeCluster}. Default = NULL.}

\item{control}{A \link[base]{list} of arguments to be passed to the optimization algorithm.
See the links for the optim parameter above. Default = NULL.}
}
\value{
A member of the S3 class \code{revenueEst}, containing the following elements: \cr \cr
Estimates, a list with elements,
\item{theta}{A vector with estimated parameters.}
\item{theta.se}{A vector with standard errors of the estimated parameters.}
\item{omegal}{A vector of firm-level labor-augmenting productivity terms,
one for each firm in each period except the first.}
\item{v_bounds}{A vector with the estimated upper and lower bound on the returns to scale,
if specified.}
\item{alpha}{A vector with estimated output elasticities, if a returns-to-scale
restriction was specified.}
\item{mu}{A vector with estimated markups (with error term), if requested and
a returns-to-scale restriction was specified.}
Model, a list with elements,
\item{theta0}{Numeric object with the optimization starting points.}
\item{thetaFS}{Numeric object with the first-step parameters.}
\item{boot.repetitions}{The number of bootstrap repetitions used for computation of the
standard errors.}
\item{elapsed.time}{Time elapsed during the estimation.}
\item{opt}{String with the optimization routine used.}
\item{opt.outcome}{Optimization outcome.}
Data, a list with elements,
\item{Y}{A vector of firms' revenue over time.}
\item{X}{A vector of state / dynamic inputs (e.g. capital).}
\item{V}{A vector of variable inputs (e.g. material inputs).}
\item{P_V}{A vector of firm- or industry-level input prices.}
\item{idvar}{A vector indexing which firm each row of the vectors corresponds to.}
\item{timevar}{A vector indexing which time period each row of the vectors corresponds to.}
}
\description{
\code{revenueCES} takes production function inputs and revenue as arguments
and estimates a CES production function with labor-augmenting productivity
up to a returns-to-scale parameter, assuming only cost minimization, as in
 \insertCite{vandijcke2022;textual}{revenueCES}.
}
\details{
Try and write
}
\examples{
Example
}
\references{
\insertRef{vandijcke2022}{revenueCES} \cr
}
