% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{systemOfEqs}
\alias{systemOfEqs}
\title{Set up system of equations}
\usage{
systemOfEqs(theta, beta)
}
\arguments{
\item{theta}{guess for parameter vector (in this case vector of exogenous variables)}

\item{beta}{calibrated parameters
let theta = (L,M,Pi,Ph,S,Y,lambda,mu,)
let beta = (beta_L, beta_M, v, sigma_CES, sigma, epsilon, N, N_h, rho_L, sigma_L, rho_M, sigma_M, rho_D, sigma_D)}
}
\description{
Takes in guess for parameter vector and spits out system of equations to be optimized
}
